global-trend-jp-publisher/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "global-trend-jp-publisher"
version = "0.1.0"
description = "Collect global trend topics and generate Japanese-ready draft posts for X and Redbook"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Yuichi" }]
dependencies = [
  "typer>=0.12.5",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "feedparser>=6.0.11",
  "requests>=2.32.3",
  "beautifulsoup4>=4.12.3",
  "langdetect>=1.0.9",
  "deep-translator>=1.11.4",
  "pillow>=10.4.0",
  "pytesseract>=0.3.13",
  "rapidocr-onnxruntime>=1.2.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "ruff>=0.6.1"
]

[project.scripts]
gtjp = "global_trend_jp_publisher.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
